discord/requirements.txt
flask[async]
selenium
python-dotenv
discord.py
nest_asyncio
requests
beautifulsoup4